Modifying layout to support titles
